menu_backend/requirements.txt
Django==1.9.8
mysqlclient==1.3.7
gunicorn==19.6.0
djangorestframework==3.4.1
